python_service/requirements.txt
# NeuroInsight CNN Service Requirements
# Install with: pip install -r requirements.txt

flask==2.3.3
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
thop>=0.1.1
numpy>=1.24.0
requests>=2.31.0